Improve performance

+++ code/analysis.py
@@ -1,5 +1,7 @@
 from utils import *
 from database import *
+from insert import *
+
 
 def load_analysis_nodes(
     nodes: pd.DataFrame, db, graph, path="./results/loading_nodes.csv"
@@ -18,39 +20,48 @@ def load_analysis_nodes(
     )
 
     qty = [100, 500, 1_000, 5_000, 10_000]
-    
-    N = 1
-    for i in tqdm(range(qty)):
-        partial_nodes = nodes.head(qty[i])
+
+    N = 10
+    for i in tqdm(range(N)):
+        # partial_nodes = nodes.head(qty[i])
+        partial_nodes = nodes
         row = {}
-        row["movie_category"] = load_movie_genres(parse_movie_generes(partial_nodes), graph)
-        row["color"] = load_colors(
-            partial_nodes["favourite_color"].dropna().unique().tolist(), graph
+        row["movie"], movies = load_movies_batch(
+            partial_nodes["favourite_movie"].dropna().unique().tolist(), db
         )
-        row["movie"] = load_movies(
-            partial_nodes["favourite_movie"].dropna().unique().tolist(), graph
+
+        row["movie_category"] = load_movie_genres_batch(
+            parse_movie_generes(partial_nodes), db
         )
-        row["university"] = load_universities(
-            partial_nodes["university"].dropna().unique().tolist(), graph
+        
+        row["color"] = load_colors_batch(
+            partial_nodes["favourite_color"].dropna().unique().tolist(), db
         )
-        row["country"] = load_countries(
+
+        row["university"], uni = load_universities_batch(
+            partial_nodes["university"].dropna().unique().tolist(), db
+        )
+        
+        row["city"], cities = load_cities_batch(parse_cities(partial_nodes), db)
+        
+        row["country"] = load_countries_batch(
             partial_nodes.groupby(["country", "continent", "country_code"])
             .size()
             .reset_index()
             .drop_duplicates(),
-            graph,
+            db
         )
-        row["city"] = load_cities(parse_cities(partial_nodes), graph)
-        row["user"] = load_users(partial_nodes, graph, db)
-        row["located_in"] = load_country_city_edges(
-            db, graph, partial_nodes[["city", "country"]].drop_duplicates()
+        
+        row["user"] = load_users_batch(partial_nodes, db, movies, uni, cities)
+        row["located_in"] = load_country_city_edges_batches(
+            db, partial_nodes[["city", "country"]].drop_duplicates()
         )
 
         times_df = pd.concat([times_df, pd.Series(row).to_frame().T], ignore_index=True)
-        if i != len(qty) - 1:
+        if i != N - 1:
             clear_all_collections(db)
 
-    times_df = times_df.set_index(pd.Index(qty))
+    # times_df = times_df.set_index(pd.Index(qty))
     times_df.to_csv(path)
 
 
@@ -70,15 +81,17 @@ def load_analysis_edges(
 
     qty = [1_000, 5_000, 10_000, 50_000, 100_000, 150_000]
 
-    for i in tqdm(range(qty)):
+    N = 10
+    for i in tqdm(range(N)):
         row = {}
-        row["likes"] = load_user_edges(edges.head(qty[i]), graph)
-        row["matches"] = load_matches(matches.head(qty[i]), graph)
+        row["likes"] = load_user_edges_batch(edges, db)
+        row["matches"] = load_matches_batch(matches, db)
 
         times_df = pd.concat([times_df, pd.Series(row).to_frame().T], ignore_index=True)
 
-        db.collections["Likes"].truncate()
-        db.collections["Matches"].truncate()
+        if i != N - 1:
+            db.collections["Likes"].truncate()
+            db.collections["Matches"].truncate()
 
-    times_df = times_df.set_index(pd.Index(qty))
+    # times_df = times_df.set_index(pd.Index(qty))
     times_df.to_csv(path)
